+++ TeamProject/src/main/java/user/service/UserServiceImpl.java
@@ -64,18 +64,28 @@ public class UserServiceImpl implements UserService {
 		Map<String, Object> map = new HashMap<String, Object>();
 		map.put("hp", hp);
 		map.put("user_id", user_id);
-		map.put("pwd", new RandomPassword().randomPassword(8));
+		map.put("pwd", this.randomPassword(8));
 		int check = userDAO.findPwCheck(map);
 		map.put("check", check);
 
 		return map;
 	}
 	
-	
-
-
-
-	
-
+	public String randomPassword (int length) {
+		int index = 0;
+		char[] charSet = new char[] {
+				'0','1','2','3','4','5','6','7','8','9'
+				,'A','B','C','D','E','F','G','H','I','J','K','L','M'
+				,'N','O','P','Q','R','S','T','U','V','W','X','Y','Z'};
+		
+			StringBuffer sb = new StringBuffer();
+			for(int i=0; i<length; i++) {
+				index = (int) (charSet.length * Math.random());
+				sb.append(charSet[index]);
+			}
+			
+			return sb.toString()+"a"+"!^";
+		
+	}
 
 }
